Reddit_sentiment_analysis/requirements.txt
data==0.4
matplotlib==3.4.3
mysql_connector_repackaged==0.3.1
nltk==3.6.3
numpy==1.21.5
pandas==1.3.3
praw==7.5.0
squarify==0.4.3
tqdm==4.62.3
